move test cases into reusable traits
