+++ src/Symfony/Bundle/FrameworkBundle/Test/WebTestCase.php
@@ -11,49 +11,13 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Test;
 
-use Symfony\Bundle\FrameworkBundle\KernelBrowser;
-use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
-
-/**
- * WebTestCase is the base class for functional tests.
- *
- * @author Fabien Potencier <fabien@symfony.com>
- */
 abstract class WebTestCase extends KernelTestCase
 {
-    use WebTestAssertionsTrait;
-
-    protected function tearDown(): void
-    {
-        parent::tearDown();
-        self::getClient(null);
-    }
-
-    /**
-     * Creates a KernelBrowser.
-     *
-     * @param array $options An array of options to pass to the createKernel method
-     * @param array $server  An array of server parameters
-     */
-    protected static function createClient(array $options = [], array $server = []): KernelBrowser
-    {
-        if (static::$booted) {
-            throw new \LogicException(\sprintf('Booting the kernel before calling "%s()" is not supported, the kernel should only be booted once.', __METHOD__));
-        }
-
-        $kernel = static::bootKernel($options);
-
-        try {
-            $client = $kernel->getContainer()->get('test.client');
-        } catch (ServiceNotFoundException) {
-            if (class_exists(KernelBrowser::class)) {
-                throw new \LogicException('You cannot create the client used in functional tests if the "framework.test" config is not set to true.');
-            }
-            throw new \LogicException('You cannot create the client used in functional tests if the BrowserKit component is not available. Try running "composer require symfony/browser-kit".');
-        }
-
-        $client->setServerParameters($server);
-
-        return self::getClient($client);
-    }
+    use Traits\BrowserKitTrait;
+    use Traits\BrowserKitAssertionsTrait;
+    use Traits\DomCrawlerTrait;
+    use Traits\DomCrawlerAssertionsTrait;
+    use Traits\HttpClientTrait;
+    use Traits\HttpClientAssertionsTrait;
+    use Traits\WebTrait;
 }
