+++ src/Symfony/Bundle/FrameworkBundle/Tests/Test/WebTestCaseTest.php
@@ -15,7 +15,6 @@ use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestAssertionsTrait;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\BrowserKit\Cookie;
 use Symfony\Component\BrowserKit\CookieJar;
@@ -388,11 +387,7 @@ class WebTestCaseTest extends TestCase
 
     private function getTester(KernelBrowser $client): WebTestCase
     {
-        $tester = new class(method_exists($this, 'name') ? $this->name() : $this->getName()) extends WebTestCase {
-            use WebTestAssertionsTrait {
-                getClient as public;
-            }
-        };
+        $tester = new class(method_exists($this, 'name') ? $this->name() : $this->getName()) extends WebTestCase {};
 
         $tester::getClient($client);
